requirements.txt
flask==2.3.3
attackcti==0.5.4
pandas==2.0.3
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
stix2==3.0.1
taxii2-client==2.3.0